requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
numpy==1.24.3
faiss-cpu==1.7.4
python-multipart==0.0.6
requests==2.31.0